minio_api/pyproject.toml
[tool.poetry]
name = "minio-api"
version = "0.1.0"
description = ""
authors = ["kooznitsa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
minio = "^7.2.7"
pyjwt = "^2.8.0"
dateutils = "^0.6.12"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"
alembic = "^1.13.1"
bcrypt = "4.0.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"
sqlalchemy-utc = "^0.14.0"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.19"
pydantic-settings = "^2.3.3"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
